Configuration moved to :config  Signed-off-by: Horacio Mijail Anton Quiles <hmijail@gmail.com>
