+++ config/src/main/java/org/hyperledger/besu/config/CrosschainConfigOptions.java
@@ -10,11 +10,11 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.hyperledger.besu.crosschain;
+package org.hyperledger.besu.config;
 
 import java.util.Map;
 
-public class CrosschainConfiguration {
+public class CrosschainConfigOptions {
   public static final String DEFAULT_PATH = "";
 
   // Crosschain feature is enabled
